consolidate theme
